+++ worker-configuration.d.ts
@@ -6,6 +6,7 @@ interface Env {
   GROQ_API_KEY: string;
   HuggingFace_API_KEY: string;
   OPEN_ROUTER_API_KEY: string;
+  FREE_OPENROUTER_API_KEY: string;
   OLLAMA_API_BASE_URL: string;
   OPENAI_LIKE_API_KEY: string;
   OPENAI_LIKE_API_BASE_URL: string;